refactor  : 기존 BoxListViewController 삭제
